third_party/blink/web_tests/http/tests/devtools/sources/debugger-step/step-through-event-listeners-expected.txt

Test that debugger will pause in all event listeners when corresponding breakpoint is set. Bug 77331.


Running: testClickBreakpoint
Script execution paused.
Call stack:
    0) listener3 (step-through-event-listeners.js:24)
    1) addListenerAndClick (step-through-event-listeners.js:33)
    2)  (:1)
Script execution resumed.
Script execution paused.
Call stack:
    0) listener1 (step-through-event-listeners.js:16)
    1) addListenerAndClick (step-through-event-listeners.js:33)
    2)  (:1)
Script execution resumed.
Script execution paused.
Call stack:
    0) listener2 (step-through-event-listeners.js:20)
    1) addListenerAndClick (step-through-event-listeners.js:33)
    2)  (:1)
Script execution resumed.
Script execution paused.
Call stack:
    0) listener3 (step-through-event-listeners.js:24)
    1) addListenerAndClick (step-through-event-listeners.js:33)
    2)  (:1)
Script execution resumed.
